requirements.txt
twilio
mysql-connector-python
secure-smtplib
uuid
json
medisearch-client
requests